Garbage

+++ src/main/java/neo/Game/Weapon.java
@@ -22,7 +22,6 @@ import neo.Game.Entity.idAnimatedEntity;
 import neo.Game.Entity.idEntity;
 import static neo.Game.Entity.signalNum_t.SIG_TOUCH;
 import neo.Game.Game.refSound_t;
-import neo.Game.GameSys.Class;
 import neo.Game.GameSys.Class.eventCallback_t;
 import neo.Game.GameSys.Class.eventCallback_t0;
 import neo.Game.GameSys.Class.eventCallback_t1;
@@ -31,7 +30,6 @@ import neo.Game.GameSys.Class.eventCallback_t4;
 import neo.Game.GameSys.Class.eventCallback_t5;
 import neo.Game.GameSys.Class.idClass;
 import neo.Game.GameSys.Class.idEventArg;
-import neo.Game.GameSys.Class.idTypeInfo;
 import neo.Game.GameSys.Event.idEventDef;
 import neo.Game.GameSys.SaveGame.idRestoreGame;
 import neo.Game.GameSys.SaveGame.idSaveGame;
@@ -2709,7 +2707,7 @@ public class Weapon {
 
                 // predict instant hit projectiles
                 if (projectileDict.GetBool("net_instanthit")) {
-                    double spreadRad = DEG2RAD(spread);
+                    float spreadRad = DEG2RAD(spread);
                     muzzle_pos = muzzleOrigin.oPlus(playerViewAxis.oGet(0).oMultiply(2.0f));
                     for (i = 0; i < num_projectiles; i++) {
                         ang = idMath.Sin(spreadRad * gameLocal.random.RandomFloat());
